fix: some valus should be updateable by global scope
https://github.com/lynx-family/lynx-stack/pull/90

This PR causes an issue:

the `registerDataProcessor` depending on the globalThis behavior.

Now we pass an updater callback to the runtime instance. This allows us to update the registerDataProcessor just like it's really assigned on globathis.

```
var {ident} = runtimeInstance
runtimeInstance.updater = () => {
 ident = runtimeInstance.storage.ident;
}
```

We just add a getter&setter for the "ident" value and call the updater on any new value set. This allows the DSL&User code to do this.

```
globalThis.ident = () => {};
ident();
```

Overall, the running code will work like this.

```
function entry(runtimeInstance) {
  var {ident} = runtimeInstance;
  runtimeInstance.updater = () => {
   ident = runtimeInstance.storage.ident;
  }
  globalThis.ident = () => {}; // this will be traped into a setter and the updater will be called.
  ident();
}
```