fix(webpack/react): `globDynamicComponentEntry` not found (#177)
<!--
  Thank you for submitting a pull request!

We appreciate the time and effort you have invested in making these
changes. Please ensure that you provide enough information to allow
others to review your pull request.

Upon submission, your pull request will be automatically assigned with
reviewers.

If you want to learn more about contributing to this project, please
visit:
https://github.com/lynx-family/lynx-stack/blob/main/CONTRIBUTING.md.
-->

## Summary

<!-- Can you explain the reasoning behind implementing this change? What
problem or issue does this pull request resolve? -->

When building a lazy bundle in a standalone project, we need the IIFE
wrapper to inject `globDynamicComponentEntry`. But we also have a `var
globDynamicComponentEntry` statement injected, which is not correct
after minification.

```js
function (globDynamicComponentEntry) {
  var globDynamicComponentEntry = globDynamicComponentEntry || '__Card__';
  // ...
  const __snapshot_xx = createSnapshot(globDynamicComponentEntry)
}
```

is minified to:

```js
function (globDynamicComponentEntry) {
  var e = e || '__Card__';
  // ...
  var x = c(e)
}
```

which result in getting **WRONG** `entryName` in lazy bundle on the
main-thread.

We just remove the `var globDynamicComponentEntry` when building
standalone lazy bundle so that the output would become:

```js
function (globDynamicComponentEntry) {
  // ...
  var x = c(globDynamicComponentEntry)
}
```

<!-- It would be helpful if you could provide any relevant context, such
as GitHub issues or related discussions. -->

## Checklist

<!--- Check and mark with an "x" -->

- [x] Tests updated (or not required).
- [x] Documentation updated (or **not required**).