fix(react): recursive hydrate in list to avoid double remove & insert when move
In the previous implementation, if a `list-item` is already rendered, and a hydrate (like reload) comes, we will remove and reinsert it. This works fine if the `diffArrayLepus` results contains no move action, but if it do, the move action will be translated to remove and insert, which may be a duplicate action.